Separate createServer
